[FIX] Increase counter size for ibf.bulk_contains.

+++ src/jaqquard_dist.cpp
@@ -61,7 +61,7 @@ void jaqquard_dist(smash_options const & options)
 
     auto worker = [&](size_t const start, size_t const end)
     {
-        auto counter = index.ibf().template counting_agent<uint16_t>();
+        auto counter = index.ibf().template counting_agent<uint32_t>();
 
         std::string result_string{};
